COMMISS.ENH: Improve orbit stability class.
Add option of letting the user define the name of  the ORM used as reference for dispersion measurements.

Let user define whether to use measured dispersion function or reference dispersion function to calculate energy evolution.

Fix a bug in calculation of energy evolution: the projection of the orbit on the dispersion subspace did not account for an offset.
Now a linear fitting is performed to extract the energy evolution.